SYSDB: Refactor the mpg and non-mpg searches out of sysdb_getgrnam() and sysdb_getgrgid() to make them more reusable
The getgrnam and getgrgid searches already special-case lookups with
overrides where in some cases the search falls back no a non-MPG search.
The upcoming special case for the hybrid mode would do something
similar, just in the opposite direction, so it makes sense to split out
the functions for just the MPG step and just the non-MPG step into
reusable functions.

Related:
https://pagure.io/SSSD/sssd/issue/3822