SYSDB: Special case getgrnam and getgrgid searches in hybrid MPG mode
In hybrid MPG mode, we want to return the MPG group only in case the
user entry has no original GID set. To achieve this, we first search
with the non-MPG filter to find 'real' groups. If that fails, we try the
MPG filter, but throw away entries that has any real GID set.

Related:
https://pagure.io/SSSD/sssd/issue/3822