PYSSS: Re-add the pysss.getgrouplist() interface
Related:
https://pagure.io/SSSD/sssd/issue/3493

Commit 0e211b8ba30c3adcdeef21ca1339b194cbfffb04 was supposed to remove
only the parts of the pysss API that relate to the local domain. But it
removed also the getgrouplist() method by accident. This method is very
important to IPA, so we need to add it back.

Reviewed-by: Alexander Bokovoy <abokovoy@redhat.com>
(cherry picked from commit f0603645f5ea5f707875807b4f815400f4b79e41)