failover: stateless mode of resolve_srv request
The fail over requests typically not only resolve a server, but also
update internal state pointers such as last_tried_server. In this
patchset, we only want to trigger DNS SRV resolution with a later patch,
therefore we add a flag to the resolve_srv_send() request which skips
touching the last_tried_server.

Note that the 'meta' server pointer still must be touched because this
is what the "_srv_" identifier expands to.