failover: Expand SRV query if needed to have at least N addresses
The administrator might choose to write up to N addresses of KDCs into
the kdcinfo files. But some setups might configure the server list like
this:
    krb5_server = kdc1, _srv_

In that case, the fail over request should make sure to also continue in
the server list and resolve the SRV query.

The code takes advantage of the restriction where only one SRV record
can be places in the server list. Therefore, the resolving can stop once
SRV result is resolved, because at that point all fo_server entries were
converted to host names.