OPENSHIFT Use the tilda API instead of tokenReview for authentication
Instead of using the TokenReview API which requires authentication and
user cannot often even review their own token, switch to using the tilda
interface which is private to OpenShift. So probably we'd have to use a
service account to authenticate sssd against the TR API.

But using the tilda interface can be unauthenticated and we're only
interested in the user who is logging in, so at least for now it's fine.