OPENSHIFT: First pass of the auth provider using the TokenReview API
Implements the OpenShift authentication provider. This requires linking
with libjansson to parse the reply from the TokenReview API and also
with libcurl to actually talk to the API with REST.

The API description can be found at:
    https://docs.okd.io/latest/rest_api/apis-authentication.k8s.io/v1.TokenReview.html
The authentication itself works by sending a POST request to the
TokenReview endpoint with the OAuth token that was passed as a password
through the PAM stack. The token is used in both the POST body (it is
the token we want reviewed) and in the Authorization header (we
authenticate using that token). The latter is problematic, because not
all users can request a review of a token, even their own. Therefore one
of the later patches uses the tilda interface instead.

The TR POST call must succeed, so in HTTP terms either 200 or 201, other
codes are converted to an error code.

The TokenReview response includes information about the user, including
their group membership and their name. The returned name is checked against
the name provided to PAM and authentication succeeds only on match.

The auth provider also adds a first option to the OpenShift provider:
    ocp_api_server_url = $API_SERVER
this option value should be set to where your Kubernetes API server
resides.

The initialization of the OpenShift provider is refactored slightly so
that there is a init_ctx that is passed to all the other providers. This
allows to set up all the providers (auth/id) at one place and then just
pass them around.