SYSDB: Allow autogenerating IDs also in the openshift domain
Because the user identities that OpenShift uses typically have no POSIX
UID or GID, we'll just be generating new entries using the
sysdb_get_new_id() that was previously only allowed by the local domain.

FIXME: maybe this should be a domain property and the function should
always generate the ID if the property was true. true/false would be set
on domain initialization.