OPENSHIFT: First pass on the identity provider
With these patches, the openshift provider can return user entries for
any request, as long as the (currently hardcoded) quota is not depleted.

After the user entry is resolved, a timer is started which checks if the
user had authenticated by the time the timer fires. If not, the user
entry is removed to not take up quota space.