OPENSHIFT: Access provider
The access providers reads a list of groups that should be allowed access
from the "ocp_allowed_groups" option. On user account PAM step, we read
the list of groups the user is a member of and only allow the user if
they are a member of at least one of the groups from the
"ocp_allowed_groups" option.